I'd have made more, but ich habe kaine LUSTgit add -u!

+++ weather_api.py
@@ -35,8 +35,9 @@ def normalize_particular_columns(df, column_names):
 
 def get_data(city_name):    
     columns_names = read_column_names('./data/header.txt')
-    # if system() is not "Linux":
-    #     sys.exit("ONLY LINUX")
+    if system() != "Linux":
+        sys.exit("Only linux is supported")
+    
     subprocess.call(shlex.split(f'./get_files.sh {city_name}'))
     df_smt = pd.read_csv(f'./data/s_m_t_{city_name}.csv', names=columns_names[1:], index_col=False)
     delete_mostly_nan_columns(df_smt)
